Fix leak of the token handle
It turns out that the WindowsIdentity constructor doesn't take
ownership of the handle that it's passed.  As a result, the handle
returned directly from DokanOpenRequestorToken never gets closed.
Constructing a subclass of SafeHandle around it ensures that it
will be closed eventually, if it's used in a using() block.  (We
do this because there are some situations where a finally{} block
won't get executed, though we also use the finally mechanism to
attempt to dispose of it as soon as we're actually done with it.)
Because handle values of both 0 and -1 are invalid as token
handles (per the Reference Source for WindowsIdentity), we need
a subclass of SafeHandleZeroOrMinusOneIsInvalid; it turns out that
SafeFileHandle is the only concrete class that derives from it.
We could implement a SafeTokenHandle class which derives from the
same base class, but that seems like a lot of work for only one
location where it's used.  So, we construct it as a SafeFileHandle
and expect that it won't be a problem.  (If Microsoft ever
revisits the SafeFileHandle code to add in a check that the
handle it's constructed with is actually a file handle, we will
need to implement SafeTokenHandle.)