Improve Buffer Handling to Reduce GC Pressure
**Background**
This is a fix for the issue discussed in #202. The default marshaller for `byte[]` arrays allocates new managed buffers for every native -> managed transition of `Read/WriteFile`. Since these
buffers are typically bigger than 85K, they are allocated on the large object heap (LOH) and requires full, expensive Gen2 garbage collection to collect. In some tests, I saw LOH allocation rates
of 500+ MB/s and 30+% CPU eaten up by GC - not good.

**The Improvement**
By manually marshalling the unmanaged buffers to managed `byte[]` arrays, we can pool/reuse the managed buffers and get away with almost no LOH allocations. This change implements that by
introducing a simple buffer pool that is utilized from `Read/WriteFile`. This is fully backwards compatible with existing applications - the full benefit is realized by using the latest
Dokan-DotNet version.

Here is a comparison of copying out a 1 GB file through DokanNetMirror before and after the change.

**Test Setup/Execution**
```
C:\DokanGCTest>fsutil file createnew 1GBfile.dat 1073741824
C:\DokanGCTest>DokanNetMirror.exe
C:\DokanGCTest>copy N:\DokanGCTest\1GBfile.dat .\copytarget.dat /Y
```

**Before**
341 full Gen2 collections to copy the 1 GB file:
![image](https://user-images.githubusercontent.com/33402265/46068250-a32b1980-c12d-11e8-9853-0ec1a06e0443.png)

**After**
0 gen2 collections (even 0 gen0 collections!) to copy the 1 GB file. You may observe that the CPU% isn't all that different between these two samples. This is because there are very few live
objects on the managed heap in this test. If Dokan was running in a larger server application with lots of live objects on the heap, the difference would be much larger.
![image](https://user-images.githubusercontent.com/33402265/46068257-a6bea080-c12d-11e8-81b5-f54b09dad318.png)

**Further Optimization**
This also opens the door to one further optimization opportunity: If the application is reading/writing the file contents from/to an unmanaged buffer (such as a memory mapped file), there is no
reason to marshall the data from unmanaged, to managed and back to unmanaged. If Dokan exposed the unmanaged `IntPtr `to the file system implementation, it can use `Buffer.MemoryCopy` to perform
the copy directly between two unmanaged buffers.

Since this requires application changes, we expose this new capability as a new sub interface of `IDokanOperations`: `IDokanOperationsUnsafe`. It introduces new overloads to `Read/WriteFile` that
takes `(IntPtr, bufferLength)` instead of `(byte[])`. The application can optionally implement this interface to take advantage of the raw buffers.

The PR also includes changes to tests to cover the new functionality:
- Changes to test infrastructure to mount two drives during testing; one for `IDokanOperations `and one for `IDokanOperationsUnsafe`. Tests select during `Setup()` which one they would like to
run against.
- A new test class for the new unsafe overloads, that reuses all the tests from `FileInfoTests `by running them in the new mode with unmanaged buffers.
- Upgraded MSTest reference versions and converted `DirectoryInfoTests `to use `DataTestMethod`/`DynamicData `instead of the old `DataSource`/XML file because these tests appeared to be flaky in
Appveyor. The new way is neater too.

This contribution from Microsoft is licensed under the MIT license.