Prevent garbage collection of delegates during DokanMain
The class DOKAN_OPERATIONS contains several delegates which are passed
to DokanMain and DokanCreateFileSystem of the C library Dokan. Because
the C library is unmanaged code, i.e. it is unaware of C#'s data
structures and garbage collection, the garbage collector must be
configured to not remove these values from memory too soon.

This commit fixes the following problem: The call to DokanMain blocks
until the file system is removed. During that time, Dokan makes repeated
calls to the delegates passed to DokanMain as DOKAN_OPERATIONS. For some
reason, the garbage collector sometimes decided during the call to
DokanMain that the delegates were no longer needed. As a result, Dokan
would make calls to already freed delegates. This would crash the
application. When debugging with an active callbackOnCollectedDelegate
MDA[1], an informative error message would be produced.

DokanCreateFileSystem has a similar problem: The delegates must remain
present even beyond the call to DokanCreateFileSystem, as it returns
immediately. In this case, the problem has already been solved by
wrapping the DOKAN_OPERATIONS in a NativeStructWrapper and returning it
to the caller. Like this, the garbage collector can't remove the
DOKAN_OPERATIONS as long as the caller keeps the returned
NativeStructWrapper around.

This commit solves the problem for DokanMain by employing exactly the
same mechanism as for DokanCreateFileSystem, avoiding duplicate code.
But since it does not return anything to the caller, it uses
GC.KeepAlive[2] to prevent the NativeStructWrapper from being garbage
collected until the call to DokanMain finishes.

[1]: https://docs.microsoft.com/en-us/dotnet/framework/debug-trace-profile/callbackoncollecteddelegate-mda
[2]: https://docs.microsoft.com/en-us/dotnet/api/system.gc.keepalive